group-7/project-1/backend/requirements.txt
flask>=2.3
flask-cors>=4.0
flask-restx>=1.3.0
python-dotenv>=1.0
dashscope>=1.14.0

# 测试依赖
pytest>=7.4
pytest-cov>=4.1
